// frontend/src/types/api/listNewslettersApiNewslettersGetSort.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * mailassist
 * OpenAPI spec version: 0.1.0
 */

export type ListNewslettersApiNewslettersGetSort = typeof ListNewslettersApiNewslettersGetSort[keyof typeof ListNewslettersApiNewslettersGetSort];


export const ListNewslettersApiNewslettersGetSort = {
  newest: 'newest',
  oldest: 'oldest',
  name: 'name',
} as const;